refactor: remove newXP names
