+++ src/features/offer/components/OfferCine/CineBlock.native.test.tsx
