+++ src/features/offer/components/OfferCine/CineBlock.tsx
