+++ src/features/offer/components/OfferCine/CineBlockSkeleton.tsx
